getStringResourceByName
- moved HashMap saving into SharedPreference
- updated getStringResourceByName
- added PixelUtil
