+++ library/build.gradle
@@ -10,8 +10,8 @@ android {
     defaultConfig {
         minSdkVersion 16
         targetSdkVersion 23
-        versionCode 4
-        versionName "0.1.3"
+        versionCode 5
+        versionName "0.1.4"
     }
     buildTypes {
         release {
