+++ library/src/main/java/com/noelchew/ncutils/PixelUtil.java
@@ -1,0 +1,23 @@
+package com.noelchew.ncutils;
+
+import android.content.Context;
+import android.util.DisplayMetrics;
+
+/**
+ * Created by noelchew on 20/09/2016.
+ */
+public class PixelUtil {
+
+    public static int dpToPx(Context context, int dp) {
+        DisplayMetrics displayMetrics = context.getResources().getDisplayMetrics();
+        int px = Math.round(dp * (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT));
+        return px;
+    }
+
+    public static int pxToDp(Context context, int px) {
+        DisplayMetrics displayMetrics = context.getResources().getDisplayMetrics();
+        int dp = Math.round(px / (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT));
+        return dp;
+    }
+
+}
