+++ sample/build.gradle
@@ -9,7 +9,7 @@ android {
         minSdkVersion 16
         targetSdkVersion 23
         versionCode 4
-        versionName "0.1.3"
+        versionName "0.1.4"
     }
     buildTypes {
         release {
